requirements.txt
glfw~=2.5.3
numpy~=1.21.6
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
Pillow~=9.2.0